Replace all uses of time.Time with custom Date class
time.Time is useful for dates, but also include a time and zone,
which is slightly worrysome when some of the code is doing
relative comparisons between dates.

To avoid accidental issues with two dates on the same day having
different zones or times by accident, replaced all uses with
a special Date class, which ensures that its internal time
is always standard, so only the date can vary.