Add pytest for etrade extraction script
I was finding it quite hard and impractical to craft pdfs
that I could submit to the repo, since they need to be
as close as possible to the real thing. I can't just
check in my own records, since they have sensitive information
in them, so I opted for a hybrid apprach.

Test materials are stored separately in a standard format,
and can be picked up by this test. The test will fail
if the user hasn't set things up correctly or at all.