Rename Decimal to DecimalOpt, and use more selectively. Fixed tests
To avoid confusion where possible, I switched uses of DecimalOpt to
raw Decimal when Null/NaN is not a valid value to have, or isn't
a case we want to have to remember to deal with.

Fixed all tests to work correctly, and started using the cmp library
in a lot of cases, because it will call .Equal and .String (usually) where provided.
This was required, because == on Decimal doesn't work, and will result
in test failures, as you can have two structs that don't compare
byte-wise, but are equivalent.