py: improve typing and error handling
The primary driver for this change was to have a clean slate for
pyright / other static analysis tools.

Where fields might be set to None, for example, they're now marked as
optional. When they're being accessed as the underling type without
None checking, I've added `mustXxx()` helper methods with an assertion
and related type annotations.

In the main loop, all of the parser results are lists, either by
wrapping the returned value (e.g. text_to_rsu_data) in a list, or by
returning a list from the type-specific helper. Instead of assigning
an "error" local that static checking thinks might be unbound, raise a
ParseError and catch it.

In the associated test, I've added an assertion for when no test
directories exist. I've also made the test more strict - if there are
unexpected errors raised when an "expected_error" file does not exist,
the test will fail.