acb: add tracking of total costs across all securities
This keeps track of the total cost basis across all securities for
every day where any trades have been made. It will also use this
information to present a yearly summary of the maximum cost of all
tracked securities in aggregate. This is useful in determining if, for
example, a T1135 form will be required for a given year.

Since it now generates multiple reports, there's a new
--csv-output-dir that will cause all of the reports to be written to
separate CSV files in the specified directory for easier consumption.

A new test, test/total_cost_test.go, checks the correctness
of the aggregate cost reporting.