app: refactor app output using new ACBWriter interface
This moves the PrintRenderTable function out of portfolio/render.go
and into a new outfmt.ACBWriter implementation.

There is currently only the outfmt.STDWriter implementation, but there
will be a CSV version in a subsequent change.

Apart from the unit tests, I also tested this against my own data and
the result was identical, with the exception of a single extra newline
after all of the output in the new version:

    --- /tmp/was.txt
    +++ /tmp/is.txt
    @@ -1266,3 +1266,4 @@
     ------------------+----------------
       Since inception | $...
     ------------------+----------------
    +