py: take an "--output-filename" arg for csv-to-xlsx, add new helper
csv-to-xlsx.py gets an `--output-filename` arg so that it doesn't have
to write to `out.xlsx`.

The new helper is useful when comparing CSV files generated with
different currency styles. Any cell that's a currency value will be
rewritten starting with a "$", no commas or spaces, and rounded to two
decimal places.