acb: handle txDelta from registered affiliate in RenderTotalCosts
These transactions have a null value for TotalAcb, and so including
them would invalidate our daily, and then yearly, totals. These should
simply be excluded.

Notes will be added to the rendered table for each instance of the
same.

Note that if there are only registered transactions for a given
security that it will not show up in the costs table at all (outside
of the warnings / notes at the bottom).