acb: handle non-default affiliate in RenderTotalCosts
When calculating the total costs, we need to also skip non-default
affiliates for unregistered affiliate transactions. This needs an
associated warning as well. Added this to the TestRenderTotalCosts
test cases.

I've adjusted the other integration tests to re-run their test cases
with and without "renderTotalCosts" as an additional smoke test.