rust: Fixes and perf improvements to questrade_statement_fmv
Fixed the month/date display, since the month column is actually the
whole date.

Major performance improvement on parsing the statement PDFs.
This PDF lib is either more thorough or just less mature than the python
pdf lib that we used to use, and so parsing the text for each page is
a lot slower for whatever reason. This change includes a couple of
changes which improve the perfrmance since the previous commit by about 10x
and brings the speed _almost_ up to parity with the python version, though it's
still something like 40% slower on my 4-thread laptop.

1) Load pages in parallel.
2) Load pages in order of expected usefulness.

Possible remaining optimization: Parallelize different statement
parsing.