rust: Add a PARALLEL_STATEMENT option to questrade_statement_fmv
I don't expect this to be used right now, but it was an interesting
experiment in performance. I tried doing both statements and pages
in parallel, but ran into blocking issues with async_std and nested
tasks. Getting the iterators to work a streams provided difficult,
and given the results, I'm happy that I didn't try to take that
to completion.