rust: Add backup pypdf PDF reader support
It pains me to do this, but I discovered that lopdf just isn't quite
ready to handle the full spectrum of shenanigans that can show up
in various PDFs. Specifically, the Morgan Stanley ETRADE trade confirmation
PDFs will just not parse at all in this library. I checked the repo, and
there are some known issues, but also still some unknown ones it
appears. The latest updates I saw in regard to that were from many
months ago, so I am not holding my breath for a fix soon.

This workaround has a build script (build.rs) install a venv into the
build output directory (just with pypdf), and adds a wrapper to run
python3 in that venv. We then can use the pdf_text.py script to
call pypdf and then split that output into pages.

All this to write the rest of the script in rust... still worth it I
think. I'll have to go back and re-decide if I should just use this lib for
all of the PDF parsing, because it's still faster to call this
subprocess than to use lopdf I think (which is also really silly).

Note: The parsed/rendered output is slightly different between the two
engines. The differences can be seen by toggling the reader with the pdf-text
binary.