Basic CsvTx support for "total amount" column and for Buy/Sell
First step for total amount specification support. This is an
alternative amount/share.

Adding this for Buy/Sell was also trivial and a good initial
run through of the logic. Changes are limited to Tx parsing
because # shares is always specified.

RoC and SFLA support is deferred because their dynamic share
amount requires changes that permiate into the Tx struct itself.