Add zip output option for CLI
This is a precursor, intended to be extended for the web app as an export option.

I'm using rawzip instead of zip specifically because it doesn't rely on
glibc or clang, since those won't allow WASM as a target. rawzip is
entirely rust, though it is a fair bit more limited.

As a result, I've opted not to use any compression, since I'd need
another library to do that. It's not super important for this use case
though.